MILESTONE1/DEMO/requirements.txt
streamlit>=1.28.0
Pillow>=10.0.0
google-generativeai>=0.3.0
pydantic>=2.0.0
numpy>=1.24.0
